refactor(registry): rename fme to feature_flags and merge knowledge-graph + semantic-layer
Rename the fme toolset to feature_flags and its 13 resource types to
feature_flag_* canonical names (with feature_flag for the flag itself), and
merge the knowledge-graph and semantic-layer toolsets into a single
software_delivery_knowledge_graph toolset (kg_* resources become
software_delivery_knowledge_graph_*; hql_query keeps its name).

All legacy toolset names and resource types continue to resolve via the alias
layer, so this is non-breaking for existing integrations and configurations.


AI-Session-Id: 6f92026d-176b-4d67-8b5d-16201370f3cb
AI-Tool: claude-code
AI-Model: unknown
